+++ api/src/main/java/com/example/api/security/UserDetailsImpl.java
@@ -16,7 +16,7 @@ public class UserDetailsImpl implements UserDetails {
     private final boolean admin;
 
     public UserDetailsImpl(Employee employee) {
-        this.id = employee.getId();
+        this.id = employee.getId().longValue();
         this.email = employee.getEmail();
         this.password = employee.getPassword();
         this.admin = employee.isAdmin(); // boolean型のgetter（例：isAdmin()）
